refactor(lemon-ui): reuse renderDetailWithLinks for error toast urls
Promotes renderDetailWithLinks from scenes/insights/EmptyStates to lib/utils and calls it from lemonToast.error, deleting the bespoke regexes, trust predicate, and "Learn more" folding. URLs now link in place with the prose untouched, so backend messages are never rewritten, the copy button keeps the full text, and both surfaces share one trust policy (isTrustedPostHogUrl).

Generated-By: PostHog Desktop
Task-Id: 61826eea-b436-49f0-a1e4-00ea58f0f5b5
