+++ frontend/src/lib/lemon-ui/LemonToast/LemonToast.tsx
@@ -6,6 +6,7 @@ import { IconCheckCircle, IconCopy, IconInfo, IconWarning, IconX } from '@postho
 
 import { getIncidentStatus, STATUS_PAGE_BASE } from 'lib/components/HelpMenu/incidentStatus'
 import { isChristmas } from 'lib/holidays'
+import { renderDetailWithLinks } from 'lib/utils/renderDetailWithLinks'
 import { hashCodeForString } from 'lib/utils/strings'
 import { writeToClipboard } from 'lib/utils/writeToClipboard'
 
@@ -161,94 +162,16 @@ function ensureToastId<T>(
     return { ...toastOptions, toastId }
 }
 
-const URL_REGEX = /https?:\/\/\S+/g
-// \S+ swallows punctuation after a URL, so sentence-ending characters are split back off the match.
-const TRAILING_PUNCTUATION_REGEX = /[.,;:!?)"']+$/
-
-function splitUrlMatch(match: string): { url: string; punctuation: string } {
-    const url = match.replace(TRAILING_PUNCTUATION_REGEX, '')
-    return { url, punctuation: match.slice(url.length) }
-}
-
-function linkifyInline(text: string): (string | JSX.Element)[] {
-    const nodes: (string | JSX.Element)[] = []
-    let cursor = 0
-    for (const match of text.matchAll(URL_REGEX)) {
-        const { url, punctuation } = splitUrlMatch(match[0])
-        if (match.index! > cursor) {
-            nodes.push(text.slice(cursor, match.index))
-        }
-        nodes.push(
-            <Link key={match.index} to={url} target="_blank">
-                {url}
-            </Link>
-        )
-        if (punctuation) {
-            nodes.push(punctuation)
-        }
-        cursor = match.index! + match[0].length
-    }
-    if (cursor < text.length) {
-        nodes.push(text.slice(cursor))
-    }
-    return nodes
-}
-
-// Only posthog.com destinations may hide behind the generic "Learn more" label. Error details can
-// carry attacker-influenced text (e.g. relayed from an integration), and a label conceals where an
-// untrusted link leads, so any other URL must stay visible as the link text.
-function isTrustedUrl(url: string): boolean {
-    try {
-        const { hostname } = new URL(url)
-        return hostname === 'posthog.com' || hostname.endsWith('.posthog.com')
-    } catch {
-        return false
-    }
-}
-
 /**
- * Backend error details often end in a raw docs URL ("... see our docs: https://posthog.com/docs/..."),
- * which as plain toast text is unclickable and wraps over several lines. A posthog.com URL ending the
- * message is folded into a "Learn more" link; any other URL becomes a link in place.
+ * Backend error details often embed a raw docs URL ("... see our docs: https://posthog.com/docs/..."),
+ * which as plain toast text is unclickable. `renderDetailWithLinks` links PostHog-host URLs in place,
+ * leaving the prose (and any untrusted URL) untouched, so the copy button still carries the full text.
  */
 export function withClickableUrls(message: string | JSX.Element): string | JSX.Element {
-    if (typeof message !== 'string') {
+    if (typeof message !== 'string' || !message.includes('http')) {
         return message
     }
-    const trimmed = message.trimEnd()
-    const matches = Array.from(trimmed.matchAll(URL_REGEX))
-    if (matches.length === 0) {
-        return message
-    }
-    const last = matches[matches.length - 1]
-    if (last.index! + last[0].length < trimmed.length) {
-        return <>{linkifyInline(trimmed)}</>
-    }
-    const { url } = splitUrlMatch(last[0])
-    if (!isTrustedUrl(url)) {
-        return <>{linkifyInline(trimmed)}</>
-    }
-    // The lead-in separator (": " or a space) reads as dangling once the URL moves into the link label,
-    // so it is dropped and the sentence closed with a period instead.
-    let lead = trimmed.slice(0, last.index).replace(/[\s:]+$/, '')
-    if (!lead) {
-        return (
-            <Link to={url} target="_blank">
-                {url}
-            </Link>
-        )
-    }
-    if (!/[.!?]$/.test(lead)) {
-        lead += '.'
-    }
-    return (
-        <>
-            {linkifyInline(lead)}{' '}
-            <Link to={url} target="_blank">
-                Learn more
-            </Link>
-        </>
-    )
+    return <>{renderDetailWithLinks(message)}</>
 }
 
 function withIncidentNote(message: string | JSX.Element): string | JSX.Element {
